# Fix a slow teardown issue